# Re-implement Serilog logs injection