--- a/tracer/src/Datadog.Trace/ClrProfiler/InstrumentationDefinitions.Generated.cs
+++ b/tracer/src/Datadog.Trace/ClrProfiler/InstrumentationDefinitions.Generated.cs
@@ -237,6 +237,10 @@ namespace Datadog.Trace.ClrProfiler
                 new("RabbitMQ.Client", "RabbitMQ.Client.Impl.ModelBase", "BasicGet",  new[] { "RabbitMQ.Client.BasicGetResult", "System.String", "System.Boolean" }, 3, 6, 9, 6, 65535, 65535, assemblyFullName, "Datadog.Trace.ClrProfiler.AutoInstrumentation.RabbitMQ.BasicGetIntegration"),
                 new("RabbitMQ.Client", "RabbitMQ.Client.Impl.ModelBase", "QueueBind",  new[] { "System.Void", "System.String", "System.String", "System.String", "System.Collections.Generic.IDictionary`2[System.String,System.Object]" }, 3, 6, 9, 6, 65535, 65535, assemblyFullName, "Datadog.Trace.ClrProfiler.AutoInstrumentation.RabbitMQ.QueueBindIntegration"),
 
+                // Serilog
+                new("Serilog", "Serilog.Core.Logger", "Dispatch",  new[] { "System.Void", "Serilog.Events.LogEvent" }, 2, 0, 0, 2, 65535, 65535, assemblyFullName, "Datadog.Trace.ClrProfiler.AutoInstrumentation.Logging.Serilog.LogsInjection.LoggerDispatchInstrumentation"),
+                new("Serilog", "Serilog.Core.Pipeline.Logger", "Dispatch",  new[] { "System.Void", "Serilog.Events.LogEvent" }, 1, 4, 214, 1, 65535, 65535, assemblyFullName, "Datadog.Trace.ClrProfiler.AutoInstrumentation.Logging.Serilog.LogsInjection.LoggerDispatchInstrumentation"),
+
                 // ServiceStackRedis
                 new("ServiceStack.Redis", "ServiceStack.Redis.RedisNativeClient", "SendReceive",  new[] { "T", "System.Byte[][]", "System.Func`1[!!0]", "System.Action`1[System.Func`1[!!0]]", "System.Boolean" }, 4, 0, 0, 5, 65535, 65535, assemblyFullName, "Datadog.Trace.ClrProfiler.AutoInstrumentation.Redis.ServiceStack.RedisNativeClientSendReceiveIntegration"),