… tracer/src/Datadog.Trace/Logging/LogsInjection/CustomSerilogLogProvider.cs
… tracer/src/Datadog.Trace/Logging/LogsInjection/SerilogEnricher.cs