## Changes

Similar to #1710 and #2096, this change modifies Serilog's logs-injection code to stop using the LibLog based subscriber to add properties directly to the `LogEvent` using automatic instrumentation.

This change has a big limitation: **Serilog logs injection now requires automatic instrumentation**.

Additionally, the CustomSerilogLogProvider type has been removed

> Once this and #2096, are merged, we will no longer need LibLog, and can remove the library entirely

## Performance Considerations

The new approach has comparable overhead compared to the lib-log implementation.

Also, worth bearing in mind, that if the `Info` level isn't enabled, we don't run with the new version, so this is a worst-case simulation (in which every log in a trace is written) compared to the lib log, (which always enriches for a trace, regardless of whether any actual logs are written or enabled).

To test the performance, I ran roughly the following benchmark code:

### Test Code
```csharp
[Benchmark]
public void WriteLogs()
{
    using (LogInjectionTracer.StartActive("Test"))
    {
        using (LogInjectionTracer.StartActive("Child"))
        {
            for (int i = 0; i < N; i++)
            {
                var logEvent = new LogEvent(
                    DateTimeOffset.Now,
                    LogEventLevel.Information,
                    exception: null,
                    new MessageTemplate("Hello", Enumerable.Empty<MessageTemplateToken>()),
                    properties: Enumerable.Empty<LogEventProperty>());

                // Simulate auto-instrumentation 

                Logger.Write(logEvent);
            }
        }
    }
}
```

### Benchmark results

The following are for `netcoreapp3.1`  - I was having issues with Sentinel killing the process with net472

|      Method |   N |       Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |---- |-----------:|----------:|----------:|-------:|------:|------:|----------:|
| None |   1 |   4.026 μs | 0.0670 μs | 0.0594 μs | 0.0305 |     - |     - |   2.06 KB |
| None |  10 |  18.712 μs | 0.1391 μs | 0.1233 μs | 0.0916 |     - |     - |   6.49 KB |
| None | 100 | 163.867 μs | 1.4601 μs | 1.2193 μs | 0.7324 |     - |     - |  50.79 KB |
|  |  | | | | | | | |
| Previous (LibLog) |   1 |   4.942 μs | 0.0459 μs | 0.0407 μs | 0.0381 |     - |     - |   2.69 KB |
| Previous (LibLog) |  10 |  29.022 μs | 0.4718 μs | 0.4414 μs | 0.1831 |     - |     - |  12.81 KB |
| Previous (LibLog) | 100 | 265.836 μs | 1.3780 μs | 1.2215 μs | 1.4648 |     - |     - | 114.04 KB |
|  |  | | | | | | | |
| New (Auto-instrument) |   1 |   5.223 μs | 0.0251 μs | 0.0235 μs | 0.0381 |     - |     - |   2.68 KB |
| New (Auto-instrument) |  10 |  30.354 μs | 0.1164 μs | 0.1089 μs | 0.1831 |     - |     - |  12.66 KB |
| New (Auto-instrument) | 100 | 279.759 μs | 1.6900 μs | 1.5808 μs | 1.4648 |     - |     - | 112.49 KB |

@DataDog/apm-dotnet